requirements.txt
firebase_admin
python-dotenv
discord
pandas